+++ CHANGELOG.md
@@ -7,10 +7,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] - TBA
+## Fixed
+- Using JSON as the medium to generate the Kubernetes Job payload had unexpected
+  issues. It caused expression errors with Sprig functions. So, now the
+  templating happens at the struct level, executing the template straight in each
+  field value
 
 ## [0.2.0] - 2023-03-03
 ### Added
-- Allow Sprig functions to be called from `JobTemplates`
+- Allow Sprig functions to be called from `JobTemplate`s
 
 ## [0.1.0] - 2023-02-15
 ### Added